:recycle: [refactor] Class Naming
